Fixed data augmentation function. Pacthes were converted to float64 instead of uint8 and int64 due to data augmentation; Pacthes are converted to float32 before being normalized fixed.

+++ utils.py
@@ -66,8 +66,8 @@ def extract_patches_mask_indices(input_image, patch_size, stride):
     return patches_array
 
 def data_augmentation(image, labels):
-    aug_imgs = np.zeros((5, image.shape[0], image.shape[1], image.shape[2]))
-    aug_lbs = np.zeros((5, image.shape[0], image.shape[1]))
+    aug_imgs = np.zeros((5, image.shape[0], image.shape[1], image.shape[2]), dtype=np.uint8)
+    aug_lbs = np.zeros((5, image.shape[0], image.shape[1]), dtype=np.int64)
 
     for i in range(0, len(aug_imgs)):
         aug_imgs[0, :, :, :] = image
